back/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2
pyjwt
Flask-HTTPAuth